backend/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
pyjwt[crypto]==2.0.1
lark-parser==0.11.1
django-filter==2.4.0
requests==2.22.0
django-anymail[sendinblue]==8.1
